requirements.txt
pandas
numpy
matplotlib
scipy
statsmodels
scikit-learn
fairlearn
pyyaml
jinja2
